feat: moves investment to its on app
