+++ core/models.py
@@ -11,10 +11,3 @@ class User(AbstractUser):
   USERNAME_FIELD = 'email'
   REQUIRED_FIELDS = []
   ...
-
-
-class Investment(models.Model):
-  owner = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.CASCADE)
-  amount = models.FloatField()
-  active = models.BooleanField(default=True, null=False)
-  created_at = models.DateTimeField()
